Revamped Chunk Basics section & a few minor edits (#14)
* Revamped Chunk Basics section & a few minor edits

I've revamped the Chunk Basics section to better match the information on the updated Chunk Data page, fixed a missing word in the "Modifying the Files (Using Community Resources)" section, and fixed the stingray_reward example in the Suit Level Codename Examples section. I would have included the full stingray_heroa_v0 and stingray_reward_islandsuit_heroa_v0 strings, but it looks like that would require a bit too much scrolling :)

I think the Chunk Basics section could still be improved further, though.

* Reverted a change in the TL;DR

Reverted a change to step 3 in the TL;DR, for brevity (being a TL;DR, and all).

* Added "the" before "Chunk Data"

I know "Chunk Data" is the name of the page, but "if we look at Chunk Data" sounds weird, so I've changed it to "if we look at the Chunk Data."

* Updated Chunk Basics

Rewrote the section on chunk hierarchy to be less ambiguous.