Edited "Background" & other small fixes (#15)
* Edited "Background" & other small fixes

Edited the "Background" section for clarity. Expanded on why SMF is preferred over RPKG Tool for generating the chunk file. Changed the adjacent chunk example to "chunk12 can not access chunk11" instead of "can not access chunk13" as the user would already be expected to know that lower numbered chunks can not access higher numbered chunks regardless of placement. Fixed one of the paths in the "Creating a Mod" section as it incorrectly pointed to SMF's `Mods` folder instead of the `Info` folder. Moved SMF's credit directly under RPKG's to match the order in which they're first mentioned.

* Small change to "This chunk file can..."